Added Service for Tenant

+++ src/main/java/com/a5lab/tabr/domain/radars/RadarRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RadarRepository extends JpaRepository<Radar, Integer> {
+public interface RadarRepository extends JpaRepository<Radar, Long> {
 }
