+++ src/main/java/com/a5lab/tabr/domain/segments/SegmentRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SegmentRepository extends JpaRepository<Segment, Integer> {
+public interface SegmentRepository extends JpaRepository<Segment, Long> {
 }
